+++ tests/MaxMind/Test/MinFraud/Model/EmailTest.php
@@ -9,10 +9,17 @@ class EmailTest extends \PHPUnit_Framework_TestCase
     public function testEmail()
     {
         $array = [
+            'first_seen' => '2017-01-02',
             'is_free' => true,
             'is_high_risk' => true
         ];
         $email = new Email($array);
+
+        $this->assertSame(
+            $array['first_seen'],
+            $email->firstSeen,
+            'firstSeen'
+        );
 
         $this->assertSame(
             $array['is_free'],